Homework 5

+++ Minibank.Core/Domains/Users/Services/UserService.cs
@@ -2,9 +2,11 @@
 using System.Threading;
 using System.Threading.Tasks;
 using FluentValidation;
+using Minibank.Core.Domains.Accounts.Repositories;
 using Minibank.Core.Domains.Dal;
 using Minibank.Core.Domains.Users.Repositories;
 using Minibank.Core.Domains.Users.Validators;
+using Minibank.Core.Exceptions.FriendlyExceptions;
 using Minibank.Core.UniversalValidators;
 
 namespace Minibank.Core.Domains.Users.Services
@@ -12,19 +14,21 @@ namespace Minibank.Core.Domains.Users.Services
     public class UserService : IUserService
     {
         private readonly IUserRepository _userRepository;
+        private readonly IBankAccountRepository _bankAccountRepository;
         private readonly IUnitOfWork _unitOfWork;
-        private readonly IValidator<User> _userValidator;
         private readonly IdEntityValidator _idValidator;
-        private readonly DeleteUserValidator _deleteUserValidator;
+        private readonly CreateUserValidator _createUserValidator;
+        private readonly UpdateUserValidator _updateUserValidator;
 
-        public UserService(IUserRepository userRepository, IUnitOfWork unitOfWork, IValidator<User> userValidator,
-            IdEntityValidator idValidator, DeleteUserValidator deleteUserValidator)
+        public UserService(IUserRepository userRepository, IBankAccountRepository bankAccountRepository, IUnitOfWork unitOfWork, IdEntityValidator idValidator,
+            CreateUserValidator createUserValidator, UpdateUserValidator updateUserValidator)
         {
             _userRepository = userRepository;
+            _bankAccountRepository = bankAccountRepository;
             _unitOfWork = unitOfWork;
-            _userValidator = userValidator;
             _idValidator = idValidator;
-            _deleteUserValidator = deleteUserValidator;
+            _createUserValidator = createUserValidator;
+            _updateUserValidator = updateUserValidator;
         }
         
         public async Task<User> GetByIdAsync(string id, CancellationToken cancellationToken)
@@ -40,25 +44,33 @@ namespace Minibank.Core.Domains.Users.Services
 
         public async Task<User> CreateUserAsync(User user, CancellationToken cancellationToken)
         {
-            await _userValidator.ValidateAndThrowAsync(user, cancellationToken);
+            await _createUserValidator.ValidateAndThrowAsync(user, cancellationToken);
             var createUser = await _userRepository.CreateUserAsync(user, cancellationToken);
             await _unitOfWork.SaveChangesAsync(cancellationToken);
             return createUser;
         }
 
-        public async Task<User> UpdateUserAsync(User user, CancellationToken cancellationToken)
+        public async Task<bool> UpdateUserAsync(User user, CancellationToken cancellationToken)
         {
-            await _idValidator.ValidateAndThrowAsync(user.Id, cancellationToken);
-            await _userValidator.ValidateAndThrowAsync(user, cancellationToken);
-            var updateUser = await _userRepository.UpdateUserAsync(user, cancellationToken);
-            await _unitOfWork.SaveChangesAsync(cancellationToken);
-            return updateUser;
+            await _updateUserValidator.ValidateAndThrowAsync(user, cancellationToken);
+            var updatingResult = await _userRepository.UpdateUserAsync(user, cancellationToken);
+            var countEntries = await _unitOfWork.SaveChangesAsync(cancellationToken);
+            return updatingResult && countEntries > 0;
         }
 
         public async Task<bool> DeleteUserAsync(string id, CancellationToken cancellationToken)
         {
-            await _deleteUserValidator.ValidateAndThrowAsync(id, cancellationToken);
-           
+            await _idValidator.ValidateAndThrowAsync(id, cancellationToken);
+            if (!await _userRepository.IsExistAsync(id, cancellationToken))
+            {
+                throw new ObjectNotFoundException($"Пользоваль с id:{id} не найден");
+            }
+            if (await _bankAccountRepository.ExistByUserIdAsync(id, cancellationToken))
+            {
+                throw new ValidationException(
+                    $"Невозможно удалить пользователя по id:{id}, так как у него имеются аккаунт(ы)");
+            }
+            
             var isDelete = await _userRepository.DeleteUserAsync(id, cancellationToken);
             var countEntries = await _unitOfWork.SaveChangesAsync(cancellationToken);
             return isDelete && countEntries > 0;
